ADFA-5083: Split tools into one file each and share the adb idioms
Refactor only - no behaviour change, and the test files are untouched, which is
what proves it.

The three list tools arrived one-per-file with pure parsing functions, but
CogoMcpServer.kt had grown to 264 lines holding server construction, tool
registration, cogo_home and is_cogo_installed. Bring the older code in line:
CogoHome.kt and CogoInstalled.kt now sit beside Projects.kt, Templates.kt and
ProjectFiles.kt, and CogoMcpServer.kt is registration only.

The duplication being collapsed is real, not anticipated. PREFS_PATH and the
preferences-read command were byte-identical in two files, the
run-as ... sh -c wrapper was hand-built in five places, and
run(listOf("shell", ...)) appeared at a dozen call sites. Those move to
CogoDevice.kt and Adb.shell().

Deliberately not collapsed: Projects.kt trims line ends while the others trim
both. That reads as duplication but the difference is load-bearing - trimEnd
protects a project name with a leading space.

Verified against emulator-5554 after the refactor, since fakes cannot prove
command construction: all six tools return what they did before, and cogo_home
still preserves an unrelated preference while writing its own exactly once.
