+++ mcp/src/main/kotlin/com/itsaky/androidide/mcp/Adb.kt
@@ -1,5 +1,8 @@
 package com.itsaky.androidide.mcp
 
+import io.modelcontextprotocol.kotlin.sdk.types.CallToolResult
+import io.modelcontextprotocol.kotlin.sdk.types.TextContent
+
 data class AdbResult(
 	val exitCode: Int,
 	val stdout: String,
@@ -8,6 +11,24 @@ data class AdbResult(
 
 fun interface Adb {
 	fun run(args: List<String>): AdbResult
+}
+
+/** Runs [args] under `adb shell`. */
+internal fun Adb.shell(vararg args: String): AdbResult = run(listOf("shell", *args))
+
+/**
+ * Turns a failed adb call into an error result.
+ *
+ * Shared by every adb-backed tool so they answer the same way: a failed call means
+ * we do not know, which is never the same as a negative answer. Collapsing the two
+ * would make an unreachable device look like a missing app, or an empty project.
+ */
+internal fun adbFailure(result: AdbResult): CallToolResult {
+	val detail = result.stderr.trim().ifEmpty { result.stdout.trim() }
+	return CallToolResult(
+		content = listOf(TextContent("adb failed (exit ${result.exitCode}): $detail")),
+		isError = true,
+	)
 }
 
 class SystemAdb(
